+++ expense.http
@@ -2,10 +2,7 @@
 GET http://localhost:3000/report/income HTTP/1.1
 
 ### find expense
-GET http://localhost:3000/report/income/b7d740dd-17e4-403c-8778-bf900a63a29a HTTP/1.1
-
-### find expense
-GET http://localhost:3000/report/expense/65d20f98-0531-4d56-82bb-cd46f835cb82 HTTP/1.1
+GET http://localhost:3000/report/expense/uuid3 HTTP/1.1
 
 ### find by id
 GET http://localhost:3000/report/income/id HTTP/1.1
@@ -15,8 +12,8 @@ POST http://localhost:3000/report/income HTTP/1.1
 content-type: application/json
 
 {
-    "amount": 3000,
-    "source": "Amazon Prime"
+    "amount": 7000,
+    "source": "Udemy"
 }
 
 ### update
@@ -24,4 +21,5 @@ PUT http://localhost:3000/report/income/123 HTTP/1.1
 Content-Type: application/json; charset=utf-8
 
 ### delete
-DELETE http://localhost:3000/report/expense/c3721d1d-097b-4fe8-b2e1-1faf59b34196
+DELETE http://localhost:3000/report/income/6670594e-8d3c-4e1e-934a-7f15d5cd35b2 HTTP/1.1
+Content-Type: application/json; charset=utf-8
