+++ src/app.service.ts
@@ -1,64 +1,8 @@
-import { Injectable } from '@nestjs/common'
-import { data, ReportType, Report } from 'src/data'
-import { v4 as uuid } from 'uuid'
+import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class AppService {
-  getAllReports(type: ReportType) {
-    return data.report.filter((report) => report.type === type)
+  getHello(): string {
+    return 'Hello World!';
   }
-
-  getReportById(type: ReportType, id: string) {
-    return data.report
-      .filter((report) => report.type === type)
-      .find((report) => report.id === id)
-  }
-
-  createReport(type: ReportType, {amount, source}: Report) {
-    const newReport = {
-      id: uuid(),
-      source,
-      amount,
-      created_at: new Date(),
-      updated_at: new Date(),
-      type
-    }
-    data.report.push(newReport)
-    return newReport
-  }
-
-  updateReportById(type: ReportType, id: string, body: Report ) {
-    const reportToUpdate = data.report
-        .filter(report => report.type === type)
-        .find(report => report.id === id)
-
-    if(!reportToUpdate) 
-      return;
-
-    const reportIndex = data.report.findIndex(report => report.id === reportToUpdate.id)
-
-    data.report[reportIndex] = {
-      ...data.report[reportIndex],
-      ...body
-    }
-
-    return data.report[reportIndex]
-  }
-
-  /**
-   * 데이터 삭제
-   * @param id: string
-   * @returns 
-   */
-  deleteReportById(id: string) {
-    const reportIndex = data.report.findIndex(report => report.id === id)
-
-    if(reportIndex == -1) 
-      return;
-
-    data.report.slice(reportIndex, 1)
-
-    return;
-  }
-
 }
